refactor: move die rolls into the existing /roll command
