+++ Server/Interactions/RollCommandGroup.cs
@@ -77,6 +77,31 @@ public class RollCommandGroup : InteractionModuleBase
         var roll = new ProgressRollRandom(Random, progressScore, description, challengeDie1, challengeDie2);
         await roll.EntityAsResponse(RespondAsync).ConfigureAwait(false);
     }
+
+    [SlashCommand("game-dice", "Rolls dice")]
+    public async Task RollDie([ComplexParameter] DieNotation first)
+    {
+        List<int> firstRollResults = GenerateRollList(first);
+        //List<int> secondRollResults = GenerateRollList(second);
+
+        var outputString = $"{first.Number}d{first.Sides}: {string.Join(", ", firstRollResults)}";
+        //if (second != null) outputString += $"\n\n{second.Number}d{second.Sides}: {string.Join(", ", secondRollResults)}";
+        await RespondAsync(outputString).ConfigureAwait(false);
+    }
+
+    private List<int> GenerateRollList(DieNotation die)
+    {
+        var rollList = new List<int>();
+        if (die != null)
+        {
+            for (var i = 0; i < die.Number; i++)
+            {
+                var roll = Random.Next(1, die.Sides + 1);
+                rollList.Add(roll);
+            }
+        }
+        return rollList;
+    }
 }
 
 public class RollInteractions : InteractionModuleBase<SocketInteractionContext<SocketMessageComponent>>
